Fix reservation booking_amount_cents different type errors

+++ app/interactions/reservations/save.rb
@@ -12,7 +12,7 @@ module Reservations
           string :state
           integer :booking_page_id, default: nil
           integer :booking_option_id, default: nil
-          integer :booking_amount_cents,  default: nil
+          decimal :booking_amount_cents,  default: nil
           string :booking_amount_currency, default: nil
           boolean :tax_include, default: nil
           time :booking_at, default: nil
